requirements.txt
Pillow
numpy
pandas
flask
keras